fix: try split
