+++ action.yml
@@ -13,8 +13,8 @@ inputs:
 
 runs:
   using: "node16"
-  main: "dist/index.js"
-  post: "dist/index.js"
+  main: "dist/main/index.js"
+  post: "dist/post/index.js"
 
 branding:
   icon: "sunrise"
