+++ src/sleep.js
@@ -1,0 +1,6 @@
+const sleepSync = (ms) => {
+    const end = new Date().getTime() + ms;
+    while (new Date().getTime() < end) { /* do nothing */ }
+}
+
+exports.sleepSync = sleepSync;
